requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.19.1
fire==0.7.1
fonttools==4.60.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pillow==11.3.0
pybboxes==0.1.6
pycocotools==2.0.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
safetensors==0.6.2
sahi==0.11.36
shapely==2.1.2
six==1.17.0
sympy==1.14.0
termcolor==3.1.0
terminaltables==3.1.10
timm==1.0.20
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
